`migrate_to_v2`: split into multiple files, stub out setting vols to read-only
